chore: move bin logic to lib
