+++ rust/frontend/Cargo.toml
@@ -21,12 +21,14 @@ pgwire = { path = "../utils/pgwire" }
 rand = "0.8.0"
 risingwave-pb = { path = "../prost" }
 risingwave_common = { path = "../common" }
+risingwave_logging = { path = "../utils/logging" }
 risingwave_meta = { path = "../meta" }
 risingwave_rpc_client = { path = "../rpc_client" }
 risingwave_source = { path = "../source" }
 risingwave_sqlparser = { path = "../sqlparser" }
 smallvec = { version = "1.6.1", features = ["serde"] }
 thiserror = "1"
+tikv-jemallocator = "0.4"
 tokio = { version = "1", features = [
     "rt",
     "rt-multi-thread",
