+++ rust/frontend/src/lib.rs
@@ -14,7 +14,12 @@ pub mod utils;
 extern crate log;
 pub mod test_utils;
 
+use std::sync::Arc;
+
 use clap::Parser;
+use pgwire::pg_server::pg_serve;
+use session::SessionManagerImpl;
+
 #[derive(Parser, Clone)]
 pub struct FrontendOpts {
     #[clap(long, default_value = "127.0.0.1:4566")]
@@ -22,4 +27,10 @@ pub struct FrontendOpts {
 
     #[clap(long, default_value = "http://127.0.0.1:5690")]
     pub meta_addr: String,
+}
+
+/// Start frontend
+pub async fn start(opts: FrontendOpts) {
+    let session_mgr = Arc::new(SessionManagerImpl::new(&opts).await.unwrap());
+    pg_serve(&opts.host, session_mgr).await.unwrap();
 }
