+++ rust/meta/Cargo.toml
@@ -34,6 +34,7 @@ prometheus = "0.13"
 prost = "0.9"
 risingwave-pb = { path = "../prost" }
 risingwave_common = { path = "../common" }
+risingwave_logging = { path = "../utils/logging" }
 risingwave_rpc_client = { path = "../rpc_client" }
 risingwave_storage = { path = "../storage" }
 serde = { version = "1", features = ["derive"] }
@@ -42,6 +43,7 @@ serde_json = "1"
 smallvec = "1"
 tempfile = "3"
 thiserror = "1"
+tikv-jemallocator = "0.4"
 tokio = { version = "1", features = [
     "rt",
     "rt-multi-thread",
