+++ rust/meta/src/lib.rs
@@ -25,3 +25,49 @@ pub mod rpc;
 pub mod storage;
 mod stream;
 pub mod test_utils;
+
+use clap::Parser;
+
+use crate::rpc::server::{rpc_serve, MetaStoreBackend};
+
+#[derive(Debug, Parser)]
+pub struct MetaNodeOpts {
+    #[clap(long, default_value = "127.0.0.1:5690")]
+    host: String,
+
+    #[clap(long)]
+    dashboard_host: Option<String>,
+
+    #[clap(long)]
+    prometheus_host: Option<String>,
+
+    #[clap(long, default_value_t = String::from(""))]
+    backend: String,
+
+    #[clap(long, default_value_t = String::from(""))]
+    etcd_endpoints: String,
+}
+
+/// Start meta node
+pub async fn start(opts: MetaNodeOpts) {
+    let addr = opts.host.parse().unwrap();
+    let dashboard_addr = opts.dashboard_host.map(|x| x.parse().unwrap());
+    let prometheus_addr = opts.prometheus_host.map(|x| x.parse().unwrap());
+    let backend = match opts.backend.as_str() {
+        "etcd" => MetaStoreBackend::Etcd {
+            endpoints: opts
+                .etcd_endpoints
+                .split(',')
+                .map(|x| x.to_string())
+                .collect(),
+        },
+        "mem" | "" => MetaStoreBackend::Mem,
+        _ => panic!("unknown backend"),
+    };
+
+    tracing::info!("Starting meta server at {}", addr);
+    let (join_handle, _shutdown_send) = rpc_serve(addr, prometheus_addr, dashboard_addr, backend)
+        .await
+        .unwrap();
+    join_handle.await.unwrap();
+}
