+++ rust/utils/logging/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "risingwave_logging"
+version = "0.1.0"
+edition = "2021"
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+async-trait = "0.1"
+futures = { version = "0.3", default-features = false, features = ["alloc"] }
+isahc = "1" # isahc is the http client used for tracing. Always set it as the same version as opentelemetry-jaeger's.
+opentelemetry = { version = "0.17", features = ["rt-tokio", "trace"] }
+opentelemetry-jaeger = { version = "0.16", features = [
+    "rt-tokio",
+    "collector_client",
+    "isahc",
+    "isahc_collector_client",
+] }
+thrift = "0.15" # thift is used for our implementation of `RwTokio` runtime. Always set it as the same version as opentelemetry-jaeger's.
+tokio = { version = "1", features = [
+    "rt",
+    "rt-multi-thread",
+    "sync",
+    "macros",
+    "time",
+    "signal",
+    "fs",
+] }
+tokio-stream = "0.1"
+tracing = { version = "0.1", features = ["release_max_level_info"] }
+tracing-opentelemetry = "0.17"
+tracing-subscriber = "0.3"
